refactor(mainnet): utility config module & tests
